Fix warning

+++ src/xtd.forms/src/xtd/forms/menu_item.cpp
@@ -173,7 +173,7 @@ void menu_item::destroy_menu_handle(intptr_t handle) {
   native::menu_item::destroy(handle);
 }
 
-int menu_item::menu_id() const {
+intptr_t menu_item::menu_id() const {
   return native::menu_item::menu_id(data_->handle_);
 }
